Move backup of auth service to separate directory
