+++ auth/api/http/backup/responses.go
@@ -1,0 +1,74 @@
+package backup
+
+import (
+	"net/http"
+	"time"
+
+	"github.com/MainfluxLabs/mainflux/pkg/apiutil"
+)
+
+var (
+	_ apiutil.Response = (*backupRes)(nil)
+	_ apiutil.Response = (*restoreRes)(nil)
+)
+
+type viewOrgRes struct {
+	ID          string         `json:"id"`
+	Name        string         `json:"name"`
+	OwnerID     string         `json:"owner_id"`
+	Description string         `json:"description,omitempty"`
+	Metadata    map[string]any `json:"metadata,omitempty"`
+	CreatedAt   time.Time      `json:"created_at"`
+	UpdatedAt   time.Time      `json:"updated_at"`
+}
+
+func (res viewOrgRes) Code() int {
+	return http.StatusOK
+}
+
+func (res viewOrgRes) Headers() map[string]string {
+	return map[string]string{}
+}
+
+func (res viewOrgRes) Empty() bool {
+	return false
+}
+
+type viewOrgMembership struct {
+	MemberID  string    `json:"member_id"`
+	OrgID     string    `json:"org_id"`
+	Role      string    `json:"role"`
+	CreatedAt time.Time `json:"created_at"`
+	UpdatedAt time.Time `json:"updated_at"`
+}
+
+type backupRes struct {
+	Orgs           []viewOrgRes        `json:"orgs"`
+	OrgMemberships []viewOrgMembership `json:"org_memberships"`
+}
+
+func (res backupRes) Code() int {
+	return http.StatusOK
+}
+
+func (res backupRes) Headers() map[string]string {
+	return map[string]string{}
+}
+
+func (res backupRes) Empty() bool {
+	return false
+}
+
+type restoreRes struct{}
+
+func (res restoreRes) Code() int {
+	return http.StatusCreated
+}
+
+func (res restoreRes) Headers() map[string]string {
+	return map[string]string{}
+}
+
+func (res restoreRes) Empty() bool {
+	return true
+}
