+++ auth/api/http/orgs/endpoint_test.go
@@ -7,7 +7,6 @@ import (
 	"io"
 	"net/http"
 	"net/http/httptest"
-	"sort"
 	"strconv"
 	"strings"
 	"testing"
@@ -854,225 +853,6 @@ func TestSearchOrgs(t *testing.T) {
 	}
 }
 
-func TestBackup(t *testing.T) {
-	svc := newService()
-	_, adminToken, err := svc.Issue(context.Background(), "", auth.Key{Type: auth.LoginKey, IssuedAt: time.Now(), IssuerID: id, Subject: email})
-	assert.Nil(t, err, fmt.Sprintf("Issuing login key expected to succeed: %s", err))
-	_, viewerToken, err := svc.Issue(context.Background(), "", auth.Key{Type: auth.LoginKey, IssuedAt: time.Now(), IssuerID: viewerID, Subject: viewerEmail})
-	assert.Nil(t, err, fmt.Sprintf("Issuing login key expected to succeed: %s", err))
-
-	ts := newServer(svc)
-	defer ts.Close()
-	client := ts.Client()
-
-	o, err := svc.CreateOrg(context.Background(), adminToken, org)
-	require.Nil(t, err, fmt.Sprintf("unexpected error: %s\n", err))
-
-	memberships := []auth.OrgMembership{viewer, editor, admin}
-	err = svc.CreateOrgMemberships(context.Background(), adminToken, o.ID, memberships...)
-	require.Nil(t, err, fmt.Sprintf("unexpected error: %s\n", err))
-
-	err = svc.AssignRole(context.Background(), id, auth.RoleAdmin)
-	require.Nil(t, err, fmt.Sprintf("unexpected error: %s\n", err))
-
-	or := []orgRes{
-		{
-			ID:          o.ID,
-			OwnerID:     o.OwnerID,
-			Name:        o.Name,
-			Description: o.Description,
-			Metadata:    o.Metadata,
-		},
-	}
-
-	m := []viewOrgMembership{
-		{
-			MemberID: id,
-			OrgID:    o.ID,
-			Role:     auth.Owner,
-		},
-		{
-			MemberID: adminID,
-			OrgID:    o.ID,
-			Role:     auth.Admin,
-		},
-		{
-			MemberID: editorID,
-			OrgID:    o.ID,
-			Role:     auth.Editor,
-		},
-		{
-			MemberID: viewerID,
-			OrgID:    o.ID,
-			Role:     auth.Viewer,
-		},
-	}
-
-	data := backup{or, m}
-
-	cases := []struct {
-		desc   string
-		token  string
-		res    backup
-		status int
-	}{
-		{
-			desc:   "backup with invalid auth token",
-			token:  wrongValue,
-			res:    backup{},
-			status: http.StatusUnauthorized,
-		},
-		{
-			desc:   "backup without auth token",
-			token:  "",
-			res:    backup{},
-			status: http.StatusUnauthorized,
-		},
-		{
-			desc:   "backup with unauthorized credentials",
-			token:  viewerToken,
-			res:    backup{},
-			status: http.StatusForbidden,
-		},
-		{
-			desc:   "backup with admin credentials",
-			token:  adminToken,
-			res:    data,
-			status: http.StatusOK,
-		},
-	}
-
-	for _, tc := range cases {
-		req := testRequest{
-			client: client,
-			method: http.MethodGet,
-			url:    fmt.Sprintf("%s/backup", ts.URL),
-			token:  tc.token,
-		}
-		res, err := req.make()
-		assert.Nil(t, err, fmt.Sprintf("%s: unexpected error %s", tc.desc, err))
-		var data backup
-		err = json.NewDecoder(res.Body).Decode(&data)
-
-		sort.Slice(data.OrgMemberships, func(i, j int) bool {
-			return data.OrgMemberships[i].MemberID < data.OrgMemberships[j].MemberID
-		})
-
-		assert.Nil(t, err, fmt.Sprintf("%s: unexpected error %s", tc.desc, err))
-		assert.Equal(t, tc.res, data, fmt.Sprintf("%s: expected body %s got %s", tc.desc, tc.res, data))
-		assert.Equal(t, tc.status, res.StatusCode, fmt.Sprintf("%s: expected status code %d got %d", tc.desc, tc.status, res.StatusCode))
-
-	}
-}
-
-func TestRestore(t *testing.T) {
-	svc := newService()
-	_, adminToken, err := svc.Issue(context.Background(), "", auth.Key{Type: auth.LoginKey, IssuedAt: time.Now(), IssuerID: id, Subject: email})
-	assert.Nil(t, err, fmt.Sprintf("Issuing login key expected to succeed: %s", err))
-	_, viewerToken, err := svc.Issue(context.Background(), "", auth.Key{Type: auth.LoginKey, IssuedAt: time.Now(), IssuerID: viewerID, Subject: viewerEmail})
-	assert.Nil(t, err, fmt.Sprintf("Issuing login key expected to succeed: %s", err))
-
-	err = svc.AssignRole(context.Background(), id, auth.RoleAdmin)
-	require.Nil(t, err, fmt.Sprintf("unexpected error: %s\n", err))
-	ts := newServer(svc)
-	defer ts.Close()
-	client := ts.Client()
-
-	orgID, err := idProvider.ID()
-	require.Nil(t, err, fmt.Sprintf("unexpected error: %s\n", err))
-
-	or := []orgRes{
-		{
-			ID:          orgID,
-			OwnerID:     id,
-			Name:        org.Name,
-			Description: org.Description,
-			Metadata:    org.Metadata,
-		},
-	}
-
-	m := []viewOrgMembership{
-		{
-			MemberID: viewerID,
-			OrgID:    orgID,
-			Role:     auth.Viewer,
-		},
-		{
-			MemberID: editorID,
-			OrgID:    orgID,
-			Role:     auth.Editor,
-		},
-		{
-			MemberID: adminID,
-			OrgID:    orgID,
-			Role:     auth.Admin,
-		},
-	}
-
-	data := toJSON(backup{
-		Orgs:           or,
-		OrgMemberships: m,
-	})
-
-	cases := []struct {
-		desc   string
-		token  string
-		req    string
-		status int
-	}{
-		{
-			desc:   "restore from backup",
-			token:  adminToken,
-			req:    data,
-			status: http.StatusCreated,
-		},
-		{
-			desc:   "restore from backup with invalid auth token",
-			token:  wrongValue,
-			req:    data,
-			status: http.StatusUnauthorized,
-		},
-		{
-			desc:   "restore from backup without auth token",
-			token:  "",
-			req:    data,
-			status: http.StatusUnauthorized,
-		},
-		{
-			desc:   "restore from backup with invalid request body",
-			token:  adminToken,
-			req:    "}",
-			status: http.StatusBadRequest,
-		},
-		{
-			desc:   "restore from backup with unauthorized credentials",
-			token:  viewerToken,
-			req:    data,
-			status: http.StatusForbidden,
-		},
-		{
-			desc:   "restore from backup without request body",
-			token:  adminToken,
-			req:    "",
-			status: http.StatusBadRequest,
-		},
-	}
-
-	for _, tc := range cases {
-		req := testRequest{
-			client: client,
-			method: http.MethodPost,
-			url:    fmt.Sprintf("%s/restore", ts.URL),
-			token:  tc.token,
-			body:   strings.NewReader(tc.req),
-		}
-		res, err := req.make()
-		assert.Nil(t, err, fmt.Sprintf("%s: unexpected error %s", tc.desc, err))
-		assert.Equal(t, tc.status, res.StatusCode, fmt.Sprintf("%s: expected status code %d got %d", tc.desc, tc.status, res.StatusCode))
-
-	}
-}
-
 type orgRes struct {
 	ID          string         `json:"id"`
 	OwnerID     string         `json:"owner_id"`
@@ -1091,14 +871,4 @@ type pageRes struct {
 	Offset uint64 `json:"offset"`
 	Total  uint64 `json:"total"`
 	Name   string `json:"name"`
-}
-
-type viewOrgMembership struct {
-	MemberID string `json:"member_id"`
-	OrgID    string `json:"org_id"`
-	Role     string `json:"role"`
-}
-type backup struct {
-	Orgs           []orgRes            `json:"orgs"`
-	OrgMemberships []viewOrgMembership `json:"org_memberships"`
 }
