+++ auth/api/http/orgs/requests.go
@@ -117,33 +117,3 @@ func (req deleteOrgsReq) validate() error {
 
 	return nil
 }
-
-type backupReq struct {
-	token string
-}
-
-func (req backupReq) validate() error {
-	if req.token == "" {
-		return apiutil.ErrBearerToken
-	}
-
-	return nil
-}
-
-type restoreReq struct {
-	token          string
-	Orgs           []viewOrgRes        `json:"orgs"`
-	OrgMemberships []viewOrgMembership `json:"org_memberships"`
-}
-
-func (req restoreReq) validate() error {
-	if req.token == "" {
-		return apiutil.ErrBearerToken
-	}
-
-	if len(req.Orgs) == 0 && len(req.OrgMemberships) == 0 {
-		return apiutil.ErrEmptyList
-	}
-
-	return nil
-}
