+++ auth/api/http/orgs/transport.go
@@ -71,20 +71,6 @@ func MakeHandler(svc auth.Service, mux *bone.Mux, tracer opentracing.Tracer, log
 		opts...,
 	))
 
-	mux.Get("/backup", kithttp.NewServer(
-		kitot.TraceServer(tracer, "backup")(backupEndpoint(svc)),
-		decodeBackup,
-		encodeResponse,
-		opts...,
-	))
-
-	mux.Post("/restore", kithttp.NewServer(
-		kitot.TraceServer(tracer, "restore")(restoreEndpoint(svc)),
-		decodeRestore,
-		encodeResponse,
-		opts...,
-	))
-
 	return mux
 }
 
@@ -160,26 +146,6 @@ func decodeDeleteOrgs(_ context.Context, r *http.Request) (any, error) {
 	}
 
 	req := deleteOrgsReq{
-		token: apiutil.ExtractBearerToken(r),
-	}
-
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		return nil, errors.Wrap(apiutil.ErrMalformedEntity, err)
-	}
-
-	return req, nil
-}
-
-func decodeBackup(_ context.Context, r *http.Request) (any, error) {
-	req := backupReq{
-		token: apiutil.ExtractBearerToken(r),
-	}
-
-	return req, nil
-}
-
-func decodeRestore(_ context.Context, r *http.Request) (any, error) {
-	req := restoreReq{
 		token: apiutil.ExtractBearerToken(r),
 	}
